+++ package.json
@@ -15,7 +15,7 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "autoprefixer": "^9.8.8",
-    "axios": "^0.24.0",
+    "axios": "^0.21.1",
     "dateformat": "^4.5.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
