+++ src/endpoints.js
@@ -2,14 +2,14 @@ import React from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
 import SprintList from './containers/SprintList';
 import TrackSprint from './components/TrackSprint';
-import SprintDetail from './containers/SprintDetail';
+import SingleSprint from './containers/SingleSprint';
 import Progress from './components/Progress';
 import UpdateSprint from './components/UpdateSprint';
 import Settings from './components/Settings';
 
 const Endpoints = () => (
   <Switch>
-    <Route path="/sprints/:sprint" exact component={SprintDetail} />
+    <Route path="/sprints/:sprint" exact component={SingleSprint} />
     <Route path="/sprints/:sprint/edit" exact component={UpdateSprint} />
     <Route path="/feed" exact component={SprintList} />
     <Route path="/" exact component={TrackSprint} />
